[Security] Implement the context switch for COOP.
This CL lands the Cross-Origin-Opener-Policy "changes to navigation",
as per: https://gist.github.com/annevk/6f2dd8c79c77123f39797f6bdac43f3e

In practice, it forces a BrowsingInstance switch when we have a COOP
mismatch. This is necessary but not sufficient, as we we want to ensure
the following JavaScript behavior:

(1) window.open() returned WindowProxy is closed when we know that we
need to switch.
(2) opener is null in opened popups.
(3) window.name is empty in opened popups.

This CL solves (2) and (3). Currently (1) is untested and will follow in
a different patch.

Bug:922191
Change-Id: I7d5169f7d53bfc28ea01a1a03c7658e0645bc492