[IntersectionObserver] Allow Document argument to constructor
This implements a behavioral change as described in:

https://github.com/w3c/IntersectionObserver/issues/372

If the 'root' argument to the IntersectionObserver constructor is a
Document, then the observer will clip to the Document's root
scroller. For the top Document, this is the same behavior as the
implicit root. For an iframe Document, this is new behavior which is
unobtainable by other means.

BUG=1015183

Change-Id: I234bfce28ab954bb15b8d157bc22ee2d2469d5e4